moved api.rs into mod.rs
